requirements.txt
elasticsearch==8.11.1
openai==0.28.1
fastapi==0.88.0
uvicorn==0.20.0
python-dotenv==0.21.1
pydantic==1.10.2
python-multipart==0.0.5
jinja2==3.1.2
httpx==0.24.1 